pyproject.toml
[project]
name = "sloy"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "oyindonesia", email = "sre-admin@oyindonesia.com" }
]
dependencies = [
    "flask>=3.0.3",
    "elasticsearch>=8.14.0",
    "pyyaml>=6.0.1",
    "dateutils>=0.6.12",
    "google-auth>=2.31.0",
    "python-dotenv>=1.0.1",
    "slack-sdk>=3.31.0",
    "gspread>=6.1.2",
    "google>=3.0.0",
    "oauth2client>=4.1.3",
    "waitress>=3.0.0",
    "google-cloud-logging>=3.11.0",
    "numpy>=2.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
dev = "sloy:main.run_app_dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sloy"]
